refactor: move `ProjectFilterSet` with it's enums from `core.filters` to `project.filters`
