+++ docker-app/qfieldcloud/project/filters.py
